Changes IPv6 prefix for MEX02
Normally we would never change an IP prefix for an existing site, but in
this case it is okay, since IPv6 never worked from day one, so there
should be no tests that have been run against this site using IPv6.
Because of this we can still gaurantee a one-to-one relationship between
experiment and IP address in the data.