Add "ipv6 " label to node blackbox_exporter targets (#209)
* Adds ipv6 label to node blackbox_exporter targets

We already add an ipv6 label to all targets generated for experiments.
This commit does the same for nodes, since there are some targets for
which we need to have this information at the node level e.g., ssh
targets.

* Updates unit tests for mlabconfig.py

Includes new 'ipv6' label for node targets, as well as a new test to be
sure that the ipv6 label gets set to "missing" when a site has no ipv6.

* Fixes the ipv6_disabled unit test for experiment

While adding an ipv6 label for node targets I realized that the unit
test for ipv6 labels for experiments was broken. The test wasn't failing
because it was written incorrectly, but now it should be working as
expected.