Removes DNS deploy check for equal RR counts
This was a very blunt check to be sure that each experiment had the
exact same number of DNS resource records. However, this falls apart
when it comes to "cloud" NDT deployments, which will only ever create
NDT RRs, causing this check to fail. This was never really a great
safeguard anyway, and doesn't work with virtual NDT nodes.