Adds new cloud NDT sites chs0t and pdx0t (#215)
* Adds virtual test node pdx0t

* Adds virtual test site pdx0c to sites.jsonnet

* Adds missing closing curly bracket to pdx0t

* Removes DNS deploy check for equal RR counts

This was a very blunt check to be sure that each experiment had the
exact same number of DNS resource records. However, this falls apart
when it comes to "cloud" NDT deployments, which will only ever create
NDT RRs, causing this check to fail. This was never really a great
safeguard anyway, and doesn't work with virtual NDT nodes.

* Changes iface for pdx0t to 'ens4'

* Updates pdx0t with location in Dalles, not LAX

* Adds new virtual/cloud test site chs0t

* Updates ASN numbers for chs0t and pdx0t