Allow for virtual NDT IPv6 sites
GCP allow external IPv6 addresses in a very small number of regions. We
have had IPv4-only virtual NDT sites for quite some time, along with
configurations in this repo to handle those cases. This commit adds the
same functionality for IPv6 sites. Namely, instead of /64 prefixes, this
change will allow a virtual site to be configured with a single /128
IPv6 address.