Adds new IPv6-enabled virtual site lax0t (#222)
* Adds new virtual site lax0t with IPv6

* Allow for virtual NDT IPv6 sites

GCP allow external IPv6 addresses in a very small number of regions. We
have had IPv4-only virtual NDT sites for quite some time, along with
configurations in this repo to handle those cases. This commit adds the
same functionality for IPv6 sites. Namely, instead of /64 prefixes, this
change will allow a virtual site to be configured with a single /128
IPv6 address.

* Uses long IPv6 addr notation for lax0t

* Sets IPv6 as exact addr, not calculated based on IPv4

For physical machines, the last hextet of an IPv6 address is equivalent
to the last octet of the same IPv4 name. However, on virtual machines
the IPv6 is what it is, as assigned by the cloud provider, so we cannot
base our IPv6 address based on the IPv6.

* Reverts back to short IPv6 notation for lax0t