Adds new provider annotation + use machine-type for type
As we start to add virtual sites to the platform there is an additional
piece of data that we would like to start recording, which is what
virtual provider the site is using (e.g., GCP, Azure, AWS, etc.). This
commit adds a new "provider" annotation, which defaults to "mlab" (bare
metal machines). This commit also changes the "type" field of the
virtual machines from "gce" to the actual GCP machine-type.