Adds 10 new production GCE virtual sites (#224)
* Adds new GCE virtual site IAD07

* Adds new GCE virtual site ORD07

* Adds new GCE virtual site SEA09

* Adds new GCE virtual site LAX07

* Adds new GCE virtual site BRU06

* Fixes AS for virtual site ORD07

* Adds new GCE virtual site LHR09

* Adds new GCE virtual site FRA07

* Adds new GCE virtual site TPE02

* Adds new GCE virtual site CGK01

* Adds new GCE virtual site SIN02

* Adds new provider annotation + use machine-type for type

As we start to add virtual sites to the platform there is an additional
piece of data that we would like to start recording, which is what
virtual provider the site is using (e.g., GCP, Azure, AWS, etc.). This
commit adds a new "provider" annotation, which defaults to "mlab" (bare
metal machines). This commit also changes the "type" field of the
virtual machines from "gce" to the actual GCP machine-type.

* Sorts annotations alphabetically