Creates a single function to return net prefixes
Rather than having two separate functions to return v4 and v6 network
prefixes, this commit creates a single function that takes a "proto"
parameter. In the case of this virtual library, it ignores the parameter
and just returns and empty string no matter what.

This commit also fixes a comment to remove an outdated reference to a
"decoration" paramter to the Machine():: function.