Creates new NetworkPrefix() function
Some format assume that every site will have an IPv4 and (possibly) IPv6
prefix, but this is only valid for physical sites. This commit
introduces a new function to both the standard and virtual site libaries
to return the prefix(es) for physical sites, and an empty string for
virtual sites.