Moves logic for deprecated formats out of libs
Some older, deprecated v1 formats are fundamentally incompatible with
virtual sites. Worksarounds have to be make to allow the deprecated
format to continue to work with existing virtual sites, as well as
physical sites. This commit moves the ugly logic to accommodate them out
of the site libraries and into the deprecated formats.