Adds a v2 version of the geo.json format
The v1 format returned a record for each site, along with a network
config for the site. However, virtual sites do not have site-wide
network configs, but instead a per-machine network config. This commit
adds a new v2 format of the geo.json format that returns a record for
each machine at a site.