Adds support for multi nodes per virtual site (#225)
* Adds new GCE virtual site IAD08 (mlab-staging)

* Initial work to support virtual sites properly

There is additional work that needs to be done, but this commit at least
has v2 formats seemingly working as expected. There will be additional
fixes to get all v1 format working, as well as to optimize general
structure and add refinements.

* Further enhancements to enabling virtual sites

This commit should resolve any issues with building v1 formats. It also
adds further enhancements to the general processing of virtual sites by
adding a new lib/site_virtual.jsonnet and sites/_default_virtual.jsonnet
files.

* Fixes unit test for site_test.jsonnet

* Adds unit testing for virtual site library

And adds a few additional tests to the regular site library.

* Fixes city name for site CPT01

* Creates new NetworkPrefix() function

Some format assume that every site will have an IPv4 and (possibly) IPv6
prefix, but this is only valid for physical sites. This commit
introduces a new function to both the standard and virtual site libaries
to return the prefix(es) for physical sites, and an empty string for
virtual sites.

* Fixes unit test to use new site.NetworkPrefix function

* Uses self referential variable instead of 'site'

* Fixes unit test for standard site lib

* Removes Machine.Network() func, expands NetworkPrefix()

Expands the functionality of the NetworkPrefix() function to accept a
'machine' parameter such that it can return meaningful values for
virtuals sites, where network configs are machine specific. Removes the
Machine.Network() function, which is now supplanted by NetworkPrefix().

* Removes Machine.Network() tests, fixes bug in virtual test

* Fixes a type in variable name

* Falls back to 1st machine in $.machines if mlab1 does not exist

* Uses std.isObject() in conditional

* Makes sure that mlab1 exists for older virtual sites

* Fixes NetworkPrefix() conditional and unit test

* Moves logic for deprecated formats out of libs

Some older, deprecated v1 formats are fundamentally incompatible with
virtual sites. Worksarounds have to be make to allow the deprecated
format to continue to work with existing virtual sites, as well as
physical sites. This commit moves the ugly logic to accommodate them out
of the site libraries and into the deprecated formats.

* Adds a v2 version of the geo.json format

The v1 format returned a record for each site, along with a network
config for the site. However, virtual sites do not have site-wide
network configs, but instead a per-machine network config. This commit
adds a new v2 format of the geo.json format that returns a record for
each machine at a site.

* Updates the IPv6 address for virtual node mlab1-lax0t

The VM was deleted and recreated and was assigned a new IPv6 prefix on
creation: https://github.com/m-lab/k8s-support/issues/666

* Removes network info from v2 version of geo.json format

As far as we know, the only thing that consumes this format is
measurementlab.net/status. We can migrate that page to be the same as it
was, but just not expose IP information for each site.

* Adds required 'machine' param to site lib tests

* Renames NetworkPrefix() to NetworkCIDR()