Uses FQDN for Name field in v2 annotations.json
The uuid-annotator uses the field Annotations.Machine to specify _only_
the machine (e.g., mlab1, mlab3, etc.) not the FQDN of the host. This
commit makes the field Name reflect the FQDN, and Annotations.Machine to
be only the machine at the site to be consitent with how the
uuid-annotator understands these terms/fields.