Makes annotations.json format a map, not array
The only thing that consumes the annotations.json format is the
uuid-annotator, as far as I know. Currently, the uuid-annotator, on
initialization, loops through every single element in the array of
machines, looking for the one that matches its current hostname. This
commit will remove the need for the loop, and will allow the
uuid-annotator do simply do something like sites[<hostname>] to get the
relevant entry.