Uses FQDN for Name field in v2 annotations.json (#233)
* Uses FQDN for Name field in v2 annotations.json

The uuid-annotator uses the field Annotations.Machine to specify _only_
the machine (e.g., mlab1, mlab3, etc.) not the FQDN of the host. This
commit makes the field Name reflect the FQDN, and Annotations.Machine to
be only the machine at the site to be consitent with how the
uuid-annotator understands these terms/fields.

* Makes annotations.json format a map, not array

The only thing that consumes the annotations.json format is the
uuid-annotator, as far as I know. Currently, the uuid-annotator, on
initialization, loops through every single element in the array of
machines, looking for the one that matches its current hostname. This
commit will remove the need for the loop, and will allow the
uuid-annotator do simply do something like sites[<hostname>] to get the
relevant entry.