Adds IPv6 addresses to several virtual sites
When we were first creating virtual sites for the cloud platform nodes
GCP did not support IPv6 in every region, Indeed, it was only supported
in 4 regions. Sometime in the past ~6m GCP apparently enabled IPv6 in
what appears to be every region. This commit adds IPv6 addresses to the
original virtual nodes that were only IPv4.