Adds additional VMs to existing sites
Google will be rolling out IST in a number of new countries. This commit
adds information for the new VMs we deployed to help absorb the extra
traffic that the IST rollouts will produce.

Additionally, this commit introduces a single, new VM in the new Google
region europe-west12 in Turin, Italy.