Adds additional VMs in regions where IST will rollout (#289)
* Retires Telecom Italia site par06

https://github.com/m-lab/ops-tracker/issues/1743

* Adds additional VMs to existing sites

Google will be rolling out IST in a number of new countries. This commit
adds information for the new VMs we deployed to help absorb the extra
traffic that the IST rollouts will produce.

Additionally, this commit introduces a single, new VM in the new Google
region europe-west12 in Turin, Italy.