Adds new virtual sites BOM05 and DFW10 + fixes IPv6 for two VMs (#305)
* Adds new virtual sites BOM05 and DFW10

The metros are overloaded due to periodic, sychronized client
integrations and/or are just too busy, causing switch discards. These
new sites should help to alleviate the problems until we can find more
permanent and long term solution.

* Updates IPv6 for mlab1-ber01 and mlab1-doh01

There appears to be some sort of bug in the interaction of Terraform and
GCP in which new VMs get an ephemeral IPv6 address, even though the
Terraform configs specify using the static IPv6 address that Terraform
creates. You can run "terraform plan" and it will tell you that it is
going to update the IPv6 address to the static one, but "terraform
apply" has no apparent affect, and when you run "terraform plan" again
it still wants to update the IPv6 address to the static one, even though
it told you it had already done this, with no apparent errors. The IPv6
stuff is very new to the Google provider, and this is likely a bug we
are hitting.