Adds new "select" test site nuq0t (#308)
The ISC used to host a full M-Lab site, but could no longer do so.
Instead they said they could host 1U with a /28. The old ISC site nuq02
was retired some time ago. This is completely a test/sandbox site for
now. Rather than the typical machines+:{} definition, this new site uses
a static machines definition with only a single mlab1 machine, as well
as a /28 IPv4 prefix which will surely have unknow behavior and/or
breakage. This is just a first start for testing in sandbox.