Support IPv4 prefixes smaller than /26 (#318)
* Removes old iupui experiments names

We haven't used experiment names like ndt-iupui, or neubot-iupui for a long
time now, and we no longer need those aliases in the experiments file, which
means we will no longer create DNS records with those names either.

* Adds support for /28 sites

The changes presuppose that there are only 2 supported prefix lengths, 26 and
28. If the prefix isn't 26, then it is presumed to be 28.

* Skips switch at site where switch==null

New /28 sites will not have a switch. In those cases, the "switch" element for
a site is set to null. If this is true, then don't attempt to include switch
information in the the switches format.

* Excludes switch RR records from /28 sites

/28 sites will not have a switch, so no switch DNS records should be created
for these sites.

* Restores cors-settings, which was accidentally rm'd in previous commit

* Adds experiment_count annotation for sites

The default value is 12, and can be overridden on a site-by-site basis. For
example, on sites using a /28 IPv4 prefix this should be set to 11, since a /28
can only support 11. This also gives us the ability to possibly even support a
/29, which would only support 3 experiments.

* Escapes comments that were previously randomly inserted text

* Adds new ExperimentCount() function to site.jsonnet

Additionally, configures _net_subnet() function to accept a new proto parameter
and enforces supported prefix lengths per protocol.

Tries to be more explicit than implicit about prefix length and experiment
count, as well as other uses of prefix length.

Adds IPv6 prefix for nuq0t (ISC site).

* Adds /32 to the list of supported v4 prefixes

Virtual sites have a /32 prefix (a single address).

* Changes bcast_offset to account for /29 prefixes

* Adds missing comma to var def. in site.jsonnet