Adds back experiments ndt.iupui and mlab.neubot
Removing them breaks mlab-ns and the changes to accommodate this and make
mlab-ns still work are too many, and involve even some bugs in GMX. It's too
much work at the moment.

One practical effect this change has on supporting "minimal" sites with /28 and
/29 IPv4 prefixes is that now /28 sites will only support up to index 10, and
/29 sites will only support mask, ndt and ndt.iupui (without revtr). Both of
these outcomes is probably just fine for v1 of minimal sites.