Adds a new function MaxRate() to the site object (#324)
* Adds a new function MaxRate() to the site object

Previously, we assumed that if a site was 1g the max-rate should be one static
thing, and that if it was 10g it should be another static thing. The
introduction of "minimal" sites that only have a single server changes this.

Minimal sites should have the max-rate set higher than "full" sites. The
max-rate values only applies to physical sites, since the number is currently
static for all virtual sites.

* Corrects conditional, removing extra 'if'

* Lowers max-rate for minimal 1g sites to 300Mbits/s