Adds new site HNL02, retroactively retires HNL01 (#331)
HNL01 was a GTT site that was only operational for about a year. It went
offline in June of 2022, but remained here in siteinfo. This commit
retroactively retires HNL01 with the date that it actually went offline.

HNL01 is being replaced by HNL02. The hardware is the same, but HNL02's circuit
is provided by Cogent and has completely different IPs than HNL01.