Don't generate switch targets for sites without a switch
Now that we are supporting minimal physical sites without a switch, we don't
want to generate Prometheus switch targets for these sites. This PR adds a
simple conditional to check whether the siteinfo value for "switch" is None
("null" in the JSON), and whether the target begins with "s1-", which would
indicate a switch target.