Don't generate switch targets for sites without a switch  (#332)
* Don't generate switch targets for sites without a switch

Now that we are supporting minimal physical sites without a switch, we don't
want to generate Prometheus switch targets for these sites. This PR adds a
simple conditional to check whether the siteinfo value for "switch" is None
("null" in the JSON), and whether the target begins with "s1-", which would
indicate a switch target.

* Fixes unit tests for select_prometheus_site_targets()

Adds a "switch" field to default sites variable. Adds a new global variable for
a site where switch=None. Edits test to be sure that this site is excluded from
switch targets.

* Checks to be sure site['switch'] exists before evaluating

Today, the calls to select_prometheus_site_targets() all filter on
"only_physical", but perhaps in the future we might want targets that aren't
physical and don't have switches, in which case site['switch'] would not exist
and might throw an error, so explicitly check for its existence before trying
to figure out if its value is None.