Changes default switch model to qfx5100-48s (#348)
The new Ziply "full" sites in Portland (PDX02), Seattle (SEA10) and Spokane
(GEG01) are going to be using Juniper QFX5100-24Q switches, and will be on
40Gbps uplinks. This commit adds the suffix "-48s" to the default switch model
for physical sites. The new Ziply sites will have a model of "qfx5100-24q". The
switch-config repository will be configured with a new template for the 24Q
model.