Reduces probability for certain VMs from 50% to 25% (#356)
* Adds new experiment "ark" with index 4

We are experimenting with the idea of hosting CAIDA Ark instances on M-Lab.

* Removes sites PDX02 and SEA10

These are two Ziply sites that were going to launch, but it has been more than
a year and there is still no movement. Remove them for now, since it's unclear
if they will ever exist. If Ziply would like to deploy them at some point, then
we can bring that back.

* Reduces probability to 25% for some virtual sites

Previously, all virtual sites that were co-located in regions/metros with a
corresponding physical site were configured with a probability of 50%. This
change further reduces the probabiliy for those same sites from 50% to 25%.

* Puts virtual site del05 back to 50% probability.